+++ src/main/java/com/stage/app/Controller/StageController.java
@@ -1,5 +1,7 @@
 package com.stage.app.Controller;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,8 +9,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import com.fasterxml.jackson.annotation.JsonCreator.Mode;
+import com.stage.app.Entity.Inscription;
 import com.stage.app.Entity.Stage;
+import com.stage.app.Repository.InscriptionRepository;
 import com.stage.app.Repository.StageRepository;
 
 import jakarta.validation.Valid;
@@ -20,6 +23,9 @@ public class StageController {
 
     @Autowired
     private StageRepository stageRepository;
+    @Autowired
+    private InscriptionRepository inscriptionRepository;
+    private String errorMessage = "";
 
     @GetMapping("/stages")
     public String getStage() {
@@ -35,12 +41,21 @@ public class StageController {
     @GetMapping("/stageList")
     public String getStageList(Model model) {
         model.addAttribute("stageList", stageRepository.findAll());
+        model.addAttribute("errorMessage", errorMessage);
+        errorMessage = "";
         return "stage/stageList";
     }
 
     @GetMapping("/stageDelete/{id}")
     public String getStageDelete(@PathVariable Integer id) {
-        stageRepository.deleteById(id);
+
+        List<Inscription> inscription = inscriptionRepository.findByStageId(id);
+
+        if (!inscription.isEmpty()) {
+            errorMessage = "Ce Stage appartient déjà à une inscription !";
+        } else {
+            stageRepository.deleteById(id);
+        }
         return "redirect:/stageList";
     }
 
@@ -51,10 +66,24 @@ public class StageController {
     }
 
     @PostMapping("/stageRegistryForm")
-    public String postStageRegistryForm(@Valid Stage stage, BindingResult result) {
+    public String postStageRegistryForm(@Valid Stage stage, BindingResult result, Model model) {
         if (result.hasErrors()) {
             return "stage/stageRegistryForm";
         }
+
+        if (stage.getAgeMin() > stage.getAgeMax()) {
+            model.addAttribute("ageError", "L'age max doit être supérieur à l'age min !");
+            return "stage/stageRegistryForm";
+        }
+
+        if (stage.getDateDeb().isAfter(stage.getDateFin())) {
+
+            model.addAttribute("dateError", "la date de début doit être avant la date de fin !");
+            return "stage/stageRegistryForm";
+        }
+
+        model.addAttribute("ageError", "");
+        model.addAttribute("dateError", "");
         stageRepository.save(stage);
         return "redirect:/";
     }
